create useNewPageInput hooks
